requirements38.txt
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
fairscale==0.4.4
filelock==3.12.0
flatbuffers==2.0.7
fonttools==4.39.3
fsspec==2023.5.0
future==0.18.3
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
huggingface-hub==0.14.1
idna==2.8
imageio==2.28.1
importlib-metadata==4.12.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==14.0.6
lit==16.0.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.5.5
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rouge-score==0.1.2
rsa==4.9
sacremoses==0.0.53
scikit-image==0.20.0
scipy==1.9.1
six==1.14.0
spicy==0.16.0
sympy==1.11.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.4
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.10.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.15.0
triton==2.0.0
typing_extensions==4.3.0
urllib3==1.25.8
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
